EDELIVERY-3897 Comment out most of the properties in domibus.properties
Use defaults for the int XA properties to prevent a NPE at Tomcat start.
Remove activeMQ.embedded.configurationFile from Tomcat default
properties as this property is special, influencing the Spring
configuration (the @PropertySource in EmbeddedActiveMQConfiguration
points to the external domibus.properties file so only that property is
considered and not its default).